created a new module named spp_registry
