+++ spp_base/__manifest__.py
@@ -15,10 +15,7 @@
         "base",
         "utm",
         "mail",
-        "g2p_registry_base",
-        "g2p_registry_individual",
-        "g2p_registry_group",
-        "g2p_registry_membership",
+        "spp_registry",
         "spp_area",
         "spp_idpass",
         "spp_idqueue",
