+++ spp_change_request/__manifest__.py
@@ -17,10 +17,7 @@
     },
     "depends": [
         "base",
-        "g2p_registry_base",
-        "g2p_registry_individual",
-        "g2p_registry_group",
-        "g2p_registry_membership",
+        "spp_registry",
         "spp_service_points",
         "spp_area",
         "spp_scan_id_document",
