+++ spp_change_request_add_children_demo/__manifest__.py
@@ -11,10 +11,7 @@
     "maintainers": ["jeremi", "gonzalesedwin1123"],
     "depends": [
         "spp_change_request",
-        "g2p_registry_base",
-        "g2p_registry_individual",
-        "g2p_registry_group",
-        "g2p_registry_membership",
+        "spp_registry",
         "spp_service_points",
     ],
     "data": [
