+++ spp_change_request_add_farmer/__manifest__.py
@@ -12,9 +12,7 @@
     "depends": [
         "spp_change_request",
         "spp_farmer_registry_base",
-        "g2p_registry_individual",
-        "g2p_registry_group",
-        "g2p_registry_membership",
+        "spp_registry",
         "spp_service_points",
         "spp_land_record",
     ],
