+++ spp_custom_fields_ui/__manifest__.py
@@ -8,7 +8,7 @@
     "license": "LGPL-3",
     "development_status": "Alpha",
     "maintainers": ["jeremi", "gonzalesedwin1123"],
-    "depends": ["base", "g2p_registry_base", "g2p_registry_membership"],
+    "depends": ["base", "spp_registry"],
     "data": [
         "views/custom_fields_ui.xml",
     ],
