+++ spp_demo/__manifest__.py
@@ -12,10 +12,7 @@
     "maintainers": ["jeremi", "gonzalesedwin1123"],
     "depends": [
         "spp_base_demo",
-        "g2p_registry_base",
-        "g2p_registry_individual",
-        "g2p_registry_group",
-        "g2p_registry_membership",
+        "spp_registry",
         "g2p_programs",
         "spp_custom_field",
         "spp_custom_field_recompute_daily",
