+++ spp_exclusion_filter/__manifest__.py
@@ -11,7 +11,7 @@
     "maintainers": ["jeremi", "gonzalesedwin1123", "reichie020212"],
     "depends": [
         "base",
-        "g2p_registry_base",
+        "spp_registry",
         "g2p_programs",
         "spp_programs",
     ],
