+++ spp_farmer_registry_dashboard/__manifest__.py
@@ -14,10 +14,7 @@
     "maintainers": ["jeremi", "gonzalesedwin1123"],
     "depends": [
         "base",
-        "g2p_registry_base",
-        "g2p_registry_individual",
-        "g2p_registry_group",
-        "g2p_registry_membership",
+        "spp_registry",
         "spp_farmer_registry_base",
         "spp_farmer_registry_demo",
         "spreadsheet_dashboard",
