--- spp_import_match/static/src/legacy/js/import_records.js
@@ -1,28 +1,0 @@
-/** @odoo-module **/
-
-import {ImportRecords} from "@base_import/import_records/import_records";
-
-import {patch} from "@web/core/utils/patch";
-
-patch(ImportRecords.prototype, {
-    importRecords() {
-        const {context, resModel} = this.env.searchModel;
-        const additionalContext = {};
-        if (resModel === "res.partner") {
-            if (context.default_is_registrant) {
-                additionalContext.default_is_registrant = context.default_is_registrant;
-            }
-            if (context.default_is_group) {
-                additionalContext.default_is_group = context.default_is_group;
-            }
-            if (context.default_kind) {
-                additionalContext.default_kind = context.default_kind;
-            }
-        }
-        this.action.doAction({
-            type: "ir.actions.client",
-            tag: "import",
-            params: {model: resModel, context, additionalContext},
-        });
-    },
-});
