+++ spp_registrant_import/__manifest__.py
@@ -11,8 +11,6 @@
     "development_status": "Alpha",
     "maintainers": ["jeremi", "gonzalesedwin1123", "nhatnm0612"],
     "depends": [
-        "base_import",
-        "web",
         "spp_base",
     ],
     "data": [
@@ -20,11 +18,7 @@
         "views/spp_area_views.xml",
         "views/spp_service_point_views.xml",
     ],
-    "assets": {
-        "web.assets_backend": [
-            "spp_registrant_import/static/src/import_records/import_records.js",
-        ],
-    },
+    "assets": {},
     "application": False,
     "installable": True,
     "auto_install": False,
