+++ spp_registrant_import/tests/test_registrant.py
@@ -166,36 +166,3 @@ class TestRegistrant(TransactionCase):
             "NGUYEN FAMILY",
             "Compute name should only compute for individual!",
         )
-
-    def test_08_get_import_templates(self):
-        import_template = self.env["res.partner"].get_import_templates()
-        self.assertEqual(
-            import_template,
-            [{"label": "Import Template for Customers", "template": "/base/static/xls/res_partner.xlsx"}],
-        )
-
-        import_template = self.env["res.partner"].with_context(default_is_registrant=True).get_import_templates()
-        self.assertEqual(
-            import_template,
-            [
-                {
-                    "label": "Import Template for Individuals",
-                    "template": "/spp_registrant_import/static/xls/individual_registry.xlsx",
-                }
-            ],
-        )
-
-        import_template = (
-            self.env["res.partner"]
-            .with_context(default_is_registrant=True, default_is_group=True)
-            .get_import_templates()
-        )
-        self.assertEqual(
-            import_template,
-            [
-                {
-                    "label": "Import Template for Groups",
-                    "template": "/spp_registrant_import/static/xls/group_registry.xlsx",
-                }
-            ],
-        )
