+++ spp_registry/static/src/import_records/import_records.js
@@ -5,7 +5,7 @@ import {ImportRecords} from "@base_import/import_records/import_records";
 import {patch} from "@web/core/utils/patch";
 
 patch(ImportRecords.prototype, {
-    importRecords() {
+    getAdditionalContext() {
         const {context, resModel} = this.env.searchModel;
         const additionalContext = {};
         if (resModel === "res.partner") {
@@ -15,7 +15,15 @@ patch(ImportRecords.prototype, {
             if (context.default_is_group) {
                 additionalContext.default_is_group = context.default_is_group;
             }
+            if (context.default_kind) {
+                additionalContext.default_kind = context.default_kind;
+            }
         }
+        return additionalContext;
+    },
+    importRecords() {
+        const {context, resModel} = this.env.searchModel;
+        const additionalContext = this.getAdditionalContext();
         this.action.doAction({
             type: "ir.actions.client",
             tag: "import",
