+++ spp_registry_group_hierarchy/__manifest__.py
@@ -13,10 +13,7 @@
     "maintainers": ["jeremi", "gonzalesedwin1123"],
     "depends": [
         "base",
-        "g2p_registry_base",
-        "g2p_registry_individual",
-        "g2p_registry_group",
-        "g2p_registry_membership",
+        "spp_registry",
     ],
     "external_dependencies": {},
     "data": [
